backend/requirements-cloudrun.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
dnspython==2.8.0
pydantic[email]==2.12.5
python-dotenv==1.2.1
PyJWT==2.11.0
bcrypt==4.1.3
pandas==3.0.1
python-multipart==0.0.22
starlette==0.37.2
certifi==2024.2.2
